DTSRD-1879: add test cases
