+++ src/test/java/uk/gov/hmcts/reform/cdapi/service/impl/CaseFlagServiceImplTest.java
@@ -98,7 +98,7 @@ class CaseFlagServiceImplTest {
         if (!(flagType == null)) {
             assertEquals(flagType, caseFlag.getFlags().get(0).getFlagDetails().get(0).getName());
         } else {
-            assertEquals(4, caseFlag.getFlags().get(0).getFlagDetails().size());
+            assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
         }
 
         verify(caseFlagRepository, times(1)).findAll(anyString());
@@ -154,11 +154,7 @@ class CaseFlagServiceImplTest {
                                                                    availableExternalFlag);
         assertNotNull(caseFlag);
         assertEquals(1, caseFlag.getFlags().size());
-        if (availableExternalFlag.equalsIgnoreCase("N")) {
-            assertEquals(4, caseFlag.getFlags().get(0).getFlagDetails().size());
-        } else {
-            assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
-        }
+        assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
         verify(caseFlagRepository, times(1)).findAll(anyString());
         caseFlag.getFlags().forEach(caseFlagObj -> {
             for (FlagDetail flagDetail : caseFlagObj.getFlagDetails()) {
@@ -167,9 +163,7 @@ class CaseFlagServiceImplTest {
                 }
                 assertNotNull(flagDetail.getDefaultStatus());
                 assertNotNull(flagDetail.getExternallyAvailable());
-                boolean externallyAvailable = (StringUtils.isNotEmpty(availableExternalFlag)
-                    && (availableExternalFlag.trim().equalsIgnoreCase("y")));
-                assertEquals(externallyAvailable, flagDetail.getExternallyAvailable());
+                assertTrue(flagDetail.getExternallyAvailable());
             }
         });
     }
@@ -200,7 +194,7 @@ class CaseFlagServiceImplTest {
         var caseFlag = caseFlagService.retrieveCaseFlagByServiceId("XXXX", "", "N", "N");
         assertNotNull(caseFlag);
         assertEquals(1, caseFlag.getFlags().size());
-        assertEquals(4, caseFlag.getFlags().get(0).getFlagDetails().size());
+        assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
         verify(caseFlagRepository, times(1)).findAll(anyString());
         caseFlag.getFlags().forEach(caseFlagObj -> {
             for (FlagDetail flagDetail : caseFlagObj.getFlagDetails()) {
@@ -269,7 +263,11 @@ class CaseFlagServiceImplTest {
             List<FlagDetail> flagDetailsList = caseFlag.getFlagDetails();
             flagDetailsList.stream().forEach(flagDetail -> {
                 boolean flagExternallyAvailable = flagDetail.getExternallyAvailable();
-                assertThat(flagExternallyAvailable, anyOf(is(externallyAvailable)));
+                if (externallyAvailable) {
+                    assertThat(flagExternallyAvailable, anyOf(is(true)));
+                } else {
+                    assertThat(flagExternallyAvailable, anyOf(is(false), is(true)));
+                }
                 assertTrue(flagDetail.getParent());
                 validateChildFlags(flagDetail.getChildFlags(), externallyAvailable);
             });
@@ -281,10 +279,12 @@ class CaseFlagServiceImplTest {
         if (flagDetails != null) {
             flagDetails.stream()
                 .forEach(flagDetail -> {
-                    if (!flagDetail.getName().equalsIgnoreCase("OTHER")) {
-                        assertThat(flagDetail.getExternallyAvailable(), anyOf(is(externallyAvailable)));
-                        validateChildFlags(flagDetail.getChildFlags(), externallyAvailable);
+                    if (externallyAvailable) {
+                        assertThat(flagDetail.getExternallyAvailable(), anyOf(is(true)));
+                    } else {
+                        assertThat(flagDetail.getExternallyAvailable(), anyOf(is(false), is(true)));
                     }
+                    validateChildFlags(flagDetail.getChildFlags(), externallyAvailable);
                 });
         }
     }
@@ -391,31 +391,7 @@ class CaseFlagServiceImplTest {
         caseFlagDtoList.add(caseFlagDto3);
         caseFlagDtoList.add(caseFlagDto4);
         caseFlagDtoList.add(caseFlagDto5);
-        caseFlagDtoList.add(createCaseFlagDto(0, 6, false, true, "COMPLEX CASE"));
-        caseFlagDtoList.add(createCaseFlagDto(0, 7, false, false, "COMPLEX CASE"));
-        caseFlagDtoList.add(createCaseFlagDto(0, 8, false, false, "CASE"));
-        caseFlagDtoList.add(createCaseFlagDto(0, 9, false, false, "PARTY"));
         return caseFlagDtoList;
-    }
-
-    private CaseFlagDto createCaseFlagDto(int categoryId,
-                                          int id,
-                                          boolean externallyAvailable,
-                                          boolean isParent,
-                                          String name) {
-        var caseFlagDto5 = new CaseFlagDto();
-        caseFlagDto5.setFlagCode("CATEGORY");
-        caseFlagDto5.setCategoryId(categoryId);
-        caseFlagDto5.setCategoryPath("CASE");
-        caseFlagDto5.setId(id);
-        caseFlagDto5.setHearingRelevant(true);
-        caseFlagDto5.setRequestReason(false);
-        caseFlagDto5.setValueEn(name);
-        caseFlagDto5.setValueCy("");
-        caseFlagDto5.setIsParent(isParent);
-        caseFlagDto5.setExternallyAvailable(externallyAvailable);
-        caseFlagDto5.setDefaultStatus("Requested");
-        return caseFlagDto5;
     }
 
     List<CaseFlagDto> getCaseFlagDtoListWithLanguageInterpreter(boolean flag) {
